backend/requirements.txt
fastapi
uvicorn
arxiv
langchain
langchain-community
langchain-core
langchain-groq
langchain-text-splitters
chromadb
pypdf
python-dotenv
sentence-transformers